Don't handle as error when receive default subscribing event
